+++ src/schemas/episode.schema.ts
@@ -1,9 +1,9 @@
 const allowedSortByProperties = ['id', 'title', 'photoSrc', 'videoSrc', 'description'];
 
 const episodeIdParamSchema = {
-    description: 'Episode ID parameter',
-    tags: ['episode'],
-    summary: 'Episode ID parameter',
+    // description: 'Episode ID parameter',
+    // tags: ['episode'],
+    // summary: 'Episode ID parameter',
     params: {
         type: 'object',
         properties: {
@@ -42,9 +42,9 @@ const episodeIdParamSchema = {
 };
 
 const episodeTitleParamSchema = {
-    description: 'Episode title parameter',
-    tags: ['episode'],
-    summary: 'Episode title parameter',
+    // description: 'Episode title parameter',
+    // tags: ['episode'],
+    // summary: 'Episode title parameter',
     params: {
         type: 'object',
         properties: {
@@ -83,9 +83,9 @@ const episodeTitleParamSchema = {
 };
 
 const episodeQuerySchema = {
-    description: 'Query episodes',
-    tags: ['episode'],
-    summary: 'Query episodes',
+    // description: 'Query episodes',
+    // tags: ['episode'],
+    // summary: 'Query episodes',
     querystring: {
         type: 'object',
         properties: {
@@ -115,42 +115,42 @@ const episodeQuerySchema = {
             },
         },
     },
-    response: {
-        200: {
-            description: 'Query executed successfully',
-            type: 'array',
-            items: {
-                type: 'object',
-                properties: {
-                    id: { type: 'integer', description: 'Episode ID' },
-                    title: { type: 'string', description: 'Episode title' },
-                    photoSrc: { type: 'string', description: 'Photo source URL' },
-                    videoSrc: { type: 'string', description: 'Video source URL' },
-                    description: { type: 'string', description: 'Episode description' },
-                    serieId: { type: 'integer', description: 'Associated serie ID' },
-                },
-            },
-        },
-        400: {
-            description: 'Bad Request',
-            type: 'object',
-            properties: {
-                error: { type: 'string', description: 'Error message' },
-            },
-        },
-        500: {
-            description: 'Internal Server Error',
-            type: 'object',
-            properties: {
-                error: { type: 'string', description: 'Error message' },
-            },
-        },
-    },
-    security: [
-        {
-            apiKey: [],
-        },
-    ],
+    // response: {
+    //     200: {
+    //         description: 'Query executed successfully',
+    //         type: 'array',
+    //         items: {
+    //             type: 'object',
+    //             properties: {
+    //                 id: { type: 'integer', description: 'Episode ID' },
+    //                 title: { type: 'string', description: 'Episode title' },
+    //                 photoSrc: { type: 'string', description: 'Photo source URL' },
+    //                 videoSrc: { type: 'string', description: 'Video source URL' },
+    //                 description: { type: 'string', description: 'Episode description' },
+    //                 serieId: { type: 'integer', description: 'Associated serie ID' },
+    //             },
+    //         },
+    //     },
+    //     400: {
+    //         description: 'Bad Request',
+    //         type: 'object',
+    //         properties: {
+    //             error: { type: 'string', description: 'Error message' },
+    //         },
+    //     },
+    //     500: {
+    //         description: 'Internal Server Error',
+    //         type: 'object',
+    //         properties: {
+    //             error: { type: 'string', description: 'Error message' },
+    //         },
+    //     },
+    // },
+    // security: [
+    //     {
+    //         apiKey: [],
+    //     },
+    // ],
 };
 
 const episodeSchemaUpdate = {
